Move default_sources and sources_to_nr_vars into montblanc.util.
